Refactor AMSR-related fetch code into `amsr` subpackage
